Refactor imports: update paths for chat-bot conversation and contact form schema files
