+++ src/app/raporteaza-incident/_constants/chat-bot-conversation.ts
